requirements.txt
cachelib==0.9.0
certifi==2022.12.7
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
colorclass==2.2.2
docopt==0.6.2
Flask==2.2.2
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pip-upgrader==1.4.15
psycopg2-binary==2.9.5
pyparsing==3.0.9
python-dotenv==1.0.1
requests==2.28.1
SQLAlchemy==1.4.43
terminaltables==3.1.10
urllib3==1.26.12
Werkzeug==2.2.2
